Hotfix CRUD Kamar

+++ app/src/main/java/com/rakhmat/setorkost/adapters/ListKamarAdapter.java
@@ -9,12 +9,14 @@ import android.support.v4.app.FragmentActivity;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
+import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
 
 import com.rakhmat.setorkost.activity.MainActivity;
+import com.rakhmat.setorkost.activity.TambahKamarActivity;
 import com.rakhmat.setorkost.activity.UbahKamarActivity;
 import com.rakhmat.setorkost.fragment.KamarFragment;
 import com.rakhmat.setorkost.model.Kamar;
@@ -29,15 +31,13 @@ import io.realm.RealmConfiguration;
 
 public class ListKamarAdapter extends RecyclerView.Adapter<ListKamarAdapter.CategoryViewHolder> {
     private Context context;
-    private Activity activity;
     private List<Kamar> kamar;
     Realm realm;
     RealmHelper realmHelper;
 
-    public ListKamarAdapter(Context context, List<Kamar> kamar, Activity activity) {
+    public ListKamarAdapter(Context context, List<Kamar> kamar) {
         this.context = context;
         this.kamar = kamar;
-        this.activity = activity;
     }
 
     @NonNull
@@ -62,18 +62,33 @@ public class ListKamarAdapter extends RecyclerView.Adapter<ListKamarAdapter.Cate
         holder.tvNomorKamar.setText(model.getNomorKamar());
         holder.tvHargaKamar.setText(model.getHargaKamar());
 
-        holder.buttonHapusKamar.setOnClickListener(new View.OnClickListener() {
+        holder.buttonHapusKamar.setOnTouchListener(new View.OnTouchListener(){
+            long then = 0;
+
             @Override
-            public void onClick(View view) {
-                realmHelper.deleteKamar(model.getId());
-                notifyDataSetChanged();
+            public boolean onTouch(View view, MotionEvent motionEvent) {
+                if(motionEvent.getAction() == MotionEvent.ACTION_DOWN){
+                    then = (Long) System.currentTimeMillis();
+                }
+                else if(motionEvent.getAction() == MotionEvent.ACTION_UP){
+                    if(((Long) System.currentTimeMillis() - then) > 2100){
+                        realmHelper.deleteKamar(model.getId());
+                        notifyDataSetChanged();
+                        return true;
+                    }
+                }
+                return false;
             }
         });
 
         holder.buttonUbahKamar.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(activity, UbahKamarActivity.class);
+                Intent intent = new Intent(context, UbahKamarActivity.class);
+                intent.putExtra("ID_KAMAR", model.getId());
+                intent.putExtra("TIPE_RUMAH", model.getTipeKamar());
+                intent.putExtra("NOMOR_KAMAR", model.getNomorKamar());
+                intent.putExtra("HARGA_KAMAR", model.getHargaKamar());
                 context.startActivity(intent);
             }
         });
